apps/v-itsm/backend/requirements.txt
# v-platform dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
cryptography>=42.0.0
structlog>=24.1.0
slowapi>=0.1.9
prometheus-client>=0.19.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
email-validator>=2.1.0
httpx>=0.25.0
aiohttp>=3.9.0

# App-specific dependencies (add here)
python-ulid>=2.2.0
apscheduler>=3.10.4

# Slack/Teams outbound providers (embedded — no runtime HTTP to v-channel-bridge)
slack-sdk>=3.27.0
# aiohttp already pinned above (used by Teams Graph/webhook provider)
